kdump high-level design document for internal review
kdump is a feature of the Linux kernel that creates crash dumps in the event
of a kernel crash. When triggered, kdump exports a memory image that can be
analyzed for the purposes of debugging and determining the cause of a crash.

This document describes how to configure and use the kdump feature in SONiC
network operating system.

Change-Id: Iecbfdd5708d0063fcd4c98694580252471dcd382